+++ crates/proto/src/rr/dnssec/rdata/nsec.rs
@@ -214,7 +214,6 @@ mod tests {
 
     #[test]
     fn test() {
-        use crate::rr::dnssec::rdata::DNSSECRecordType;
         use crate::rr::RecordType;
         use std::str::FromStr;